requirements.txt
alembic
eralchemy2
openpyxl
pandas
psycopg2
python-docx
python-dotenv
requests
SQLAlchemy
